added ConnectionManager.kt
